+++ tests/unit/RouterTest.php
@@ -9,6 +9,7 @@ use Craft\Contracts\RoutesCollectionInterface;
 use Craft\Contracts\UriInterface;
 use Craft\Http\Exceptions\NotFoundHttpException;
 use Craft\Http\Route\Route;
+use Craft\Http\Route\RouteParamsParser;
 use Craft\Http\Route\Router;
 use PHPUnit\Framework\MockObject\Exception;
 use PHPUnit\Framework\TestCase;
@@ -16,13 +17,14 @@ use stdClass;
 
 class RouterTest extends TestCase
 {
-    private function createRouter($container, $routesCollection, $request)
+    private function createRouter($container, $routesCollection, $request, $response, $routeParamsParser)
     {
         return new Router(
             $container,
             $routesCollection,
             $request,
-            $this->createMock(ResponseInterface::class)
+            $response,
+            $routeParamsParser
         );
     }
 
@@ -30,6 +32,9 @@ class RouterTest extends TestCase
     {
         $request = $this->createMock(RequestInterface::class);
         $request->method('getMethod')->willReturn('GET');
+
+        $routeParamsParser = $this->createMock(RouteParamsParser::class);
+        $routeParamsParser->method('buildRegexFromRoute')->willReturn('#^/test$#');
 
         $uri = $this->createMock(UriInterface::class);
         $uri->method('getPath')->willReturn('/test');
@@ -55,7 +60,7 @@ class RouterTest extends TestCase
 
         $response = $this->createMock(ResponseInterface::class);
 
-        $router = $this->createRouter($container, $routesCollection, $request);
+        $router = $this->createRouter($container, $routesCollection, $request, $response, $routeParamsParser);
 
         $router->dispatch();
     }
@@ -73,7 +78,13 @@ class RouterTest extends TestCase
         $routesCollection = $this->createMock(RoutesCollectionInterface::class);
         $routesCollection->method('getRoutes')->willReturn([]);
 
-        $router = $this->createRouter($this->createMock(DIContainer::class), $routesCollection, $request);
+        $router = $this->createRouter(
+            $this->createMock(DIContainer::class),
+            $routesCollection,
+            $request,
+            $this->createMock(ResponseInterface::class),
+            $this->createMock(RouteParamsParser::class)
+        );
 
         $this->expectException(NotFoundHttpException::class);