+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 - Изменено:
   - Переименован класс QueryBuilder на MySql\Connection
+- Исправлено:
+  - Уменьшил уровень вложенности в AbstractCommand
 
 ## [ [0.4.2](https://github.com/C-o-d-eCraft/framework/releases/tag/0.4.2) ] - 20.10.2024
 - Реализовано: