+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Переименован класс NotAuthorizedHttpException в UnauthorizedHttpException
 - Исправлено:
   - Уменьшил уровень вложенности в AbstractCommand
+  - Вынес методы из ConsoleKernel в Input
 
 ## [ [0.4.2](https://github.com/C-o-d-eCraft/framework/releases/tag/0.4.2) ] - 20.10.2024
 - Реализовано: