+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - Уменьшил уровень вложенности в AbstractCommand
   - Вынес методы из ConsoleKernel в Input
   - Привел параметры конструктора к единообразию
+  - Удален излишний абстрактный класс Message
 
 ## [ [0.4.2](https://github.com/C-o-d-eCraft/framework/releases/tag/0.4.2) ] - 20.10.2024
 - Реализовано: