+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Формат основан на [стандарте формата CHANGELOG](https://keepachangelog.com/en/1.0.0/),
 и придерживается [правил версионирования](https://semver.org/spec/v2.0.0.html).
 
+## [ [0.4.3](https://github.com/C-o-d-eCraft/framework/releases/tag/0.4.3) ] - 07.11.2024
 - Изменено:
   - Валидация в роутере с помощью Validator 
   - Переименован класс QueryBuilder на MySql\Connection